3DLINE: corrections in member types of dxf_3dline_struct.

+++ src/3dline.h
@@ -1,7 +1,7 @@
 /*!
  * \file 3dline.h
  *
- * \author Copyright (C) 2015, 2016, 2017, 2018
+ * \author Copyright (C) 2015, 2016, 2017, 2018, 2019
  * by Bert Timmerman <bert.timmerman@xs4all.nl>.
  *
  * \brief Header file for a DXF 3D line entity (\c 3DLINE).
@@ -99,20 +99,20 @@ dxf_3dline_struct
                  * </ol>
                  * Group code = 60.\n
                  * \since Introduced in version R13. */
-        int color;
+        int16_t color;
                 /*!< Color of the entity.\n
                  * Defaults to \c BYLAYER if ommitted in the DXF file.\n
                  * Note that entities encapsulated in a block with the
                  * color \c BYBLOCK are represented in the "native" color of
                  * the \c BLOCK entity.\n
                  * Group code = 62. */
-        int paperspace;
+        int16_t paperspace;
                 /*!< Entities are to be drawn on either \c PAPERSPACE or
                  * \c MODELSPACE.\n
                  * Optional, defaults to \c DXF_MODELSPACE (0).\n
                  * Group code = 67.\n
                  * \since Introduced in version R13. */
-        int graphics_data_size;
+        int32_t graphics_data_size;
                 /*!< Number of bytes in the proxy entity graphics
                  * represented in the sub-sequent 310 groups, which are
                  * binary chunk records (optional).\n
@@ -236,12 +236,12 @@ double dxf_3dline_get_linetype_scale (Dxf3dline *line);
 Dxf3dline *dxf_3dline_set_linetype_scale (Dxf3dline *line, double linetype_scale);
 int16_t dxf_3dline_get_visibility (Dxf3dline *line);
 Dxf3dline *dxf_3dline_set_visibility (Dxf3dline *line, int16_t visibility);
-int dxf_3dline_get_color (Dxf3dline *line);
-Dxf3dline *dxf_3dline_set_color (Dxf3dline *line, int color);
-int dxf_3dline_get_paperspace (Dxf3dline *line);
-Dxf3dline *dxf_3dline_set_paperspace (Dxf3dline *line, int paperspace);
-int dxf_3dline_get_graphics_data_size (Dxf3dline *line);
-Dxf3dline *dxf_3dline_set_graphics_data_size (Dxf3dline *line, int graphics_data_size);
+int16_t dxf_3dline_get_color (Dxf3dline *line);
+Dxf3dline *dxf_3dline_set_color (Dxf3dline *line, int16_t color);
+int16_t dxf_3dline_get_paperspace (Dxf3dline *line);
+Dxf3dline *dxf_3dline_set_paperspace (Dxf3dline *line, int16_t paperspace);
+int32_t dxf_3dline_get_graphics_data_size (Dxf3dline *line);
+Dxf3dline *dxf_3dline_set_graphics_data_size (Dxf3dline *line, int32_t graphics_data_size);
 int16_t dxf_3dline_get_shadow_mode (Dxf3dline *line);
 Dxf3dline *dxf_3dline_set_shadow_mode (Dxf3dline *line, int16_t shadow_mode);
 DxfBinaryGraphicsData *dxf_3dline_get_binary_graphics_data (Dxf3dline *line);
